requirements.txt
pandas
numpy
scikit-learn
gradio
streamlit
groq
python-dotenv